requirements.txt
adafruit-ampy==1.0.7
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
bandit==1.6.2
beautifulsoup4==4.8.2
black==19.10b0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
dota2==0.3.3
ecdsa==0.15
emoji==0.5.4
ephem==3.7.7.1
esptool==2.8
fusepy==3.0.1
future==0.18.2
gevent==1.4.0
gevent-eventemitter==2.1
gitdb==4.0.4
GitPython==3.1.1
GosuGamers-API==0.2
greenlet==0.4.15
idna==2.8
iso8601==0.1.12
isort==4.3.21
lazy-object-proxy==1.4.3
liquipediapy==2.1.0
lxml==4.5.0
mccabe==0.6.1
more-itertools==8.2.0
mpy-utils==0.1.12
numpy==1.18.4
packaging==20.1
pathspec==0.7.0
pbr==5.4.5
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyaes==1.6.1
pycparser==2.19
pyjson5==1.4.7
pylint==2.4.4
pyparsing==2.4.6
pyserial==3.4
PySocks==1.7.1
pytest==5.3.5
python-dotenv==0.10.3
python-telegram-bot==12.8
pytz==2020.1
PyYAML==5.3
regex==2020.2.18
requests==2.22.0
six==1.13.0
smmap==3.0.2
soupsieve==2.0
steam==0.9.1
stevedore==1.32.0
toml==0.10.0
tornado==6.0.3
typed-ast==1.4.1
urllib3==1.25.8
vdf==3.2
wcwidth==0.1.8
wrapt==1.11.2
